new function autoSize
function autoSize can be called from within the Slot-item, for example: 
if(this.$parent.autoSize) this.$parent.autoSize();
i don't know much About your components, i'm sure there will be an easier /better way to do this, but this is working for me

why: i use it to arrange multiple statistics. of course they can be bigger or smaller, depends on configuration and data, so i never now how much space is needed. now i can call autosize of loading is finished and the grid-item will Change ist size