Emit @resized whenever pixel width/height changes
Until now, @resized was only emitted when the user resized the grid
items.

Not if e.g. the user widens the window, which causes w and h to stay
the same, but width grows.

With this commit, all changes to GridItem-s that cause width and
height to change cause @resized to be emitted.